requirements.txt
django
google-generativeai
python-dotenv
pillow
dj-database-url
psycopg2-binary
openpyxl
django-storages
google-cloud-storage
gunicorn
